LayoutTests/fast/forms/validation-messages-expected.txt

Required text input: 
Required checkbox input: 
Required radio input: 
Required radio input:   
Required file input: 
Required email input: 
Required url input: 
Required input with pattern: 
Required input with minlength=100: 
Required range with min=5: 
Required range with max=5: 
Required range with step=3 / min=0: 
Tests the HTML form validation messages being shown on UI side.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS validationBubbleContents.message is "Fill out this field"
PASS validationBubbleContents.message is "Select this checkbox"
PASS validationBubbleContents.message is "Select one of these options"
PASS validationBubbleContents.message is "Select a file"
PASS validationBubbleContents.message is "Enter an email address"
PASS validationBubbleContents.message is "Enter a URL"
PASS validationBubbleContents.message is "Match the requested format"
PASS validationBubbleContents.message is "Use at least 100 characters"
PASS validationBubbleContents.message is "Value must be greater than or equal to 5"
PASS validationBubbleContents.message is "Value must be less than or equal to 5"
PASS validationBubbleContents.message is "Enter a valid value"
PASS successfullyParsed is true

TEST COMPLETE
